+++ index.html
@@ -10,7 +10,7 @@
 
 <body>
 
-	<script src="script.js"></script>
+	<script type="module" src="script.js"></script>
 
 </body>
