+++ modules/dictionary.js
@@ -1,0 +1,74 @@
+// Создаем массив кнопок клавиатуры
+
+const englishKeys = [
+	{ value: "`", shiftValue: "~", code: "Backquote" },
+	{ value: "1", shiftValue: "!", code: "Digit1" },
+	{ value: "2", shiftValue: "@", code: "Digit2" },
+	{ value: "3", shiftValue: "#", code: "Digit3" },
+	{ value: "4", shiftValue: "$", code: "Digit4" },
+	{ value: "5", shiftValue: "%", code: "Digit5" },
+	{ value: "6", shiftValue: "^", code: "Digit6" },
+	{ value: "7", shiftValue: "&", code: "Digit7" },
+	{ value: "8", shiftValue: "*", code: "Digit8" },
+	{ value: "9", shiftValue: "(", code: "Digit9" },
+	{ value: "0", shiftValue: ")", code: "Digit0" },
+	{ value: "-", shiftValue: "_", code: "Minus" },
+	{ value: "=", shiftValue: "+", code: "Equal" },
+	{ value: "Del", shiftValue: "Delete", code: "Delete" },
+
+	{ value: "Tab", shiftValue: "Tab", code: "Tab" },
+	{ value: "Q", shiftValue: "q", code: "KeyQ" },
+	{ value: "W", shiftValue: "w", code: "KeyW" },
+	{ value: "E", shiftValue: "e", code: "KeyE" },
+	{ value: "R", shiftValue: "r", code: "KeyR" },
+	{ value: "T", shiftValue: "t", code: "KeyT" },
+	{ value: "Y", shiftValue: "y", code: "KeyY" },
+	{ value: "U", shiftValue: "u", code: "KeyU" },
+	{ value: "I", shiftValue: "i", code: "KeyI" },
+	{ value: "O", shiftValue: "o", code: "KeyO" },
+	{ value: "P", shiftValue: "p", code: "KeyP" },
+	{ value: "[", shiftValue: "{", code: "BracketLeft" },
+	{ value: "]", shiftValue: "}", code: "BracketRight" },
+	{ value: "\\", shiftValue: "|", code: "Backslash" },
+
+	{ value: "CapsLock", shiftValue: "CapsLock", code: "CapsLock" },
+	{ value: "A", shiftValue: "a", code: "KeyA" },
+	{ value: "S", shiftValue: "s", code: "KeyS" },
+	{ value: "D", shiftValue: "d", code: "KeyD" },
+	{ value: "F", shiftValue: "f", code: "KeyF" },
+	{ value: "G", shiftValue: "g", code: "KeyG" },
+	{ value: "H", shiftValue: "h", code: "KeyH" },
+	{ value: "J", shiftValue: "j", code: "KeyJ" },
+	{ value: "K", shiftValue: "k", code: "KeyK" },
+	{ value: "L", shiftValue: "l", code: "KeyL" },
+	{ value: ";", shiftValue: ":", code: "Semicolon" },
+	{ value: '"', shiftValue: "'", code: "Quote" },
+	{ value: "Enter", shiftValue: "Enter", code: "Enter" },
+
+	{ value: "Shift", shiftValue: "Shift", code: "ShiftLeft" },
+	{ value: "Z", shiftValue: "z", code: "KeyZ" },
+	{ value: "X", shiftValue: "x", code: "KeyX" },
+	{ value: "C", shiftValue: "c", code: "KeyC" },
+	{ value: "V", shiftValue: "v", code: "KeyV" },
+	{ value: "B", shiftValue: "b", code: "KeyB" },
+	{ value: "N", shiftValue: "n", code: "KeyN" },
+	{ value: "M", shiftValue: "m", code: "KeyM" },
+	{ value: ",", code: "Comma" },
+	{ value: ".", code: "Period" },
+	{ value: "/", code: "Slash" },
+	{ value: "Shift", shiftValue: "Shift", code: "ShiftRight" },
+
+	{ value: "fn", code: "Fn" },
+	{ value: "control", shiftValue: "Control", code: "ControlLeft" },
+	{ value: "option", code: "AltLeft" },
+	{ value: "⌘", code: "MetaLeft" },
+	{ value: "Space", code: "Space" },
+	{ value: "⌘", code: "MetaRight" },
+	{ value: "option", code: "AltRight" },
+	{ value: "◀", shiftValue: "◀", code: "ArrowLeft" },
+	{ value: "▲", shiftValue: "▲", code: "ArrowUp" },
+	{ value: "▼", shiftValue: "▼", code: "ArrowDown" },
+	{ value: "►", shiftValue: "►", code: "ArrowRight" }
+];
+
+export default englishKeys;
